#109: make the `WebPFree()` function part of the official interface
The `WebPFree()` function is provided in a hacky way via the `Makefile.vc` for Windows DLL builds only. There is no function signature available in a header file so it cannot be used directly from C++. And the current solution does not allow to use it on cross platform code.

This PR moves the `WebPFree()` out of the `Makefile.vc` file into one of the official header files so it can be used in cross platform code and without hacks.